fixed git repo check for version
App now uses the local version it has unless X
amount of time has passed to check for a new
version.
This is done to prevent going over the
github api limit.
The app is not actively checking for a new version.
It only checks when the app is opened or a new
page is opened.